Move JSExport protocols into separate files
