+++ Stark/API/Event.swift
@@ -9,18 +9,7 @@
 import AppKit
 import JavaScriptCore
 
-@objc
-protocol EventJSExport: JSExport {
-    init(event: String, callback: JSValue)
-
-    var name: String { get }
-
-    func disable()
-}
-
 public class Event: Handler, EventJSExport, HashableJSExport {
-    /// Initializers
-
     public required init(event: String, callback: JSValue) {
         name = event
 
@@ -41,15 +30,11 @@ public class Event: Handler, EventJSExport, HashableJSExport {
         disable()
     }
 
-    /// Instance Variables
-
     private var notification: String
 
     private var notificationCenter: NotificationCenter
 
     public var name: String
-
-    /// Instance Functions
 
     public func disable() {
         notificationCenter.removeObserver(self, name: NSNotification.Name(rawValue: notification), object: nil)
