+++ Stark/API/Timer.swift
@@ -9,16 +9,7 @@
 import AppKit
 import JavaScriptCore
 
-@objc
-protocol TimerJSExport: JSExport {
-    init(interval: TimeInterval, repeats: Bool, callback: JSValue)
-
-    func stop()
-}
-
 public class Timer: Handler, TimerJSExport, HashableJSExport {
-    /// Initializers
-
     public required init(interval: TimeInterval, repeats: Bool, callback: JSValue) {
         super.init()
 
@@ -31,11 +22,7 @@ public class Timer: Handler, TimerJSExport, HashableJSExport {
         manageCallback(callback)
     }
 
-    /// Instance Variables
-
     private var timer: Foundation.Timer?
-
-    /// Instance Functions
 
     public func stop() {
         timer?.invalidate()
